move nudge + pulsar components to pulsar dir and add deprecation warning
