+++ docs/documentation/components/pulsar.mdx
@@ -1,10 +1,12 @@
+> Deprecation note: The `Nudge` component is deprecated and will be renamed `Pulsar` in the next major version.
+
 # Usage
 
 Pulsars are best suited for situations where you need to draw attention to a part in the UI, and want to do it contextually and gently without too much of a distraction or removal from the main experience of using the app.
 
 Specifically, a pulsar is optional - the user can simply ignore the pulsing dot and continue using the app as they would. If they happen across it, the pulsar can provide a helpful tip on hover or can even guide them to their next action.
 
-It is also important to note that pulsars are not meant to be a permanent part of the UI. They should be used judiciously and with purpose. If pulsars are there all the time, and in every part of the app, they will lose its meaning.
+It is also important to note that pulsars are not meant to be a permanent part of the UI. They should be used judiciously and with purpose. If pulsars are there all the time, and in every part of the app, they will lose their meaning.
 
 ```jsx
 <Pane padding={40} display="flex" flexDirection="column" gap={32} alignItems="flex-start">
