+++ src/index.js
@@ -26,7 +26,7 @@ export { IconWrapper as Icon } from './icons/src/IconWrapper'
 export { Image } from './image'
 export { Pane, Card } from './layers'
 export { Menu } from './menu'
-export { Pulsar, Nudge } from './nudge'
+export { Pulsar, Nudge } from './pulsar'
 export { Overlay } from './overlay'
 export { Popover } from './popover'
 export { Portal } from './portal'
